COMP: Handle optional dependencies in CMakeLists.txt after modules are loaded
When this module is built as a remote module within VTK, it
is not possible to dynamically update the DAG based on
availability of the optional VTK modules it should depend on.

To workaround this limitation, we perform the following tests:

(1) testing for the availability of the target:

     - This always works for a stand-alone build because the targets
        were imported in VTKConfig.

     - but for a remote build where VTKConfig is only generated at
       the end of the configuration, it will work only for the
       optional dependencies that **happen to be enabled** before
       this module.

(2) testing for Module_${optional-depends}: This is useful only
    for the remote build and allow to detect the case where the user
    explicitly enable a module.

Ideally, the module API should be extended to support an
OPTIONAL_DEPENDS property.